refactor: extract types to a separate file
