+++ src/index.ts
@@ -1,16 +1,9 @@
+import { DatabaseItem, TestOutput, Order, User } from './types';
 import {
   calculateTotalLengthOfFullNames,
   countNumberOfUniqueCities,
 } from './utils';
 
-type DatabaseItem = {
-  type: string;
-  id: number;
-  firstName?: string;
-  lastName?: string;
-  price?: number;
-  city?: string;
-};
 
 const unorganisedListOfDatabaseItems: DatabaseItem[] = [
   {
@@ -108,32 +101,7 @@ const unorganisedListOfDatabaseItems: DatabaseItem[] = [
   },
 ];
 
-type Order = {
-  id: number;
-  price: number;
-};
 
-export type Address = {
-  id: number;
-  city: string;
-};
-
-export type User = {
-  id: number;
-  firstName: string;
-  lastName: string;
-};
-
-type TestOutput = {
-  countOfDatabaseItems: number;
-  orders: Order[];
-  totalRevenue: number;
-  addresses: Address[];
-  countOfUniqueCities: number;
-  users: User[];
-  totalLengthOfFullNames: number;
-  allFullNames: string[];
-};
 
 export default function main(): TestOutput {
   const testOutput: TestOutput = {
